requirements.txt
streamlit>=1.27.0
pandas>=2.0.0
plotly>=5.14.0
pyarrow>=20.0.0
streamlit-authenticator>=0.2.3
pyyaml>=6.0
bcrypt>=4.0.0
cryptography>=42.0.0 